pyproject.toml
[tool.poetry]
name = "sim_rl"
version = "0.1.4"
description = "Simulation Driven RL Package Utilized to Optimize Queueing Systems"
authors = ["Fatima Alani, Jinyan Wang, Jevon Charles, Joshua Forday, Aaron Ong, Vinayak Modi <fatima.al-ani23@imperial.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
torch = "2.2.0"
numpy = "1.26.4"
pandas = "2.2.0"
queueing_tool = "1.2.5"
matplotlib = "3.8.3"
wandb = "0.16.3"
PyYAML = "6.0.1"
ray = "2.9.2"
tqdm = "4.57.0"
scipy = "1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
black = "^24.3.0"
Flask = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"